// adopt-me-front/src/api/commons/api-exception/ApiException.ts
export enum ApiException {
    AUTHENTICATION_EXCEPTION = "AUTHENTICATION_EXCEPTION",
    USER_ACCOUNT_NOT_ACTIVATED_EXCEPTION = "USER_ACCOUNT_NOT_ACTIVATED_EXCEPTION",
    USER_ALREADY_EXISTS = "USER_ALREADY_EXISTS",
    GMAIL_AUTHENTICATION_EXCEPTION = "GMAIL_AUTHENTICATION_EXCEPTION",
    UNPRIVILEGED_GMAIL_DOMAIN_EXCEPTION = "UNPRIVILEGED_GMAIL_DOMAIN_EXCEPTION",
    ACTIVATE_ACCOUNT_TOKEN_ALREADY_USED_EXCEPTION = "ACTIVATE_ACCOUNT_TOKEN_ALREADY_USED_EXCEPTION",
    PASSWORD_RESET_TOKEN_EXPIRED_EXCEPTION = "PASSWORD_RESET_TOKEN_EXPIRED_EXCEPTION",
    PASSWORD_RESET_TOKEN_ALREADY_USED_EXCEPTION = "PASSWORD_RESET_TOKEN_ALREADY_USED_EXCEPTION",
    PASSWORD_RESET_GMAIL_AUTH_EXCEPTION = "PASSWORD_RESET_GMAIL_AUTH_EXCEPTION",
    GMAIL_LOGIN_PASS_AUTH_DISABLED_EXCEPTION = "GMAIL_LOGIN_PASS_AUTH_DISABLED_EXCEPTION",
    USER_NOT_FOUND_BY_EMAIL_EXCEPTION = "USER_NOT_FOUND_BY_EMAIL_EXCEPTION",
    TAG_ALREADY_EXISTS_EXCEPTION = "TAG_ALREADY_EXISTS_EXCEPTION"
}